Add Instagram comments scraping functionality to Apify webhook handler
