+++ lib/apify/handleApifyWebhook.ts
@@ -15,6 +15,7 @@ import getAccountEmails from "../supabase/accountEmails/getAccountEmails";
 import sendApifyWebhookEmail from "@/lib/apify/sendApifyWebhookEmail";
 import normalizeProfileUrl from "@/lib/utils/normalizeProfileUrl";
 import uploadLinkToArweave from "@/lib/arweave/uploadLinkToArweave";
+import runInstagramCommentsScraper from "@/lib/apify/runInstagramCommentsScraper";
 
 /**
  * Handles the Apify webhook payload: fetches dataset, saves posts, saves socials, and returns results.
@@ -31,6 +32,7 @@ export default async function handleApifyWebhook(
   let accountArtistIds: Tables<"account_artist_ids">[] = [];
   let accountEmails: Tables<"account_emails">[] = [];
   let sentEmails: unknown = null;
+  let commentsScraperResult: unknown = null;
   let dataset;
   if (datasetId) {
     try {
@@ -87,6 +89,19 @@ export default async function handleApifyWebhook(
             firstResult,
             emails.map((e) => e.email).filter(Boolean) as string[]
           );
+
+          // Trigger comment scraping for the new posts
+          if (firstResult.latestPosts && firstResult.latestPosts.length > 0) {
+            const postUrls = (firstResult.latestPosts as ApifyInstagramPost[])
+              .map((post) => post.url)
+              .filter(Boolean);
+            
+            if (postUrls.length > 0) {
+              console.log("Triggering comment scraping for posts:", postUrls);
+              commentsScraperResult = await runInstagramCommentsScraper(postUrls);
+              console.log("Comments scraper result:", commentsScraperResult);
+            }
+          }
         }
       }
     } catch (e) {
@@ -101,5 +116,6 @@ export default async function handleApifyWebhook(
     accountArtistIds,
     accountEmails,
     sentEmails,
+    commentsScraperResult,
   };
 }
